kboot: Use unsigned long long.
For the 64-bit platforms, this is a nop. Currently kboot only supports
64-bit platforms, though. If we support 32-bit in the future, this will
become important.

Noticed by:		rpokala
Sponsored by:		Netflix

+++ stand/kboot/util.c
@@ -35,7 +35,7 @@ file2str(const char *fn, char *buffer, size_t buflen)
 bool
 file2u64(const char *fn, uint64_t *val)
 {
-	unsigned long v;
+	unsigned long long v;
 	char buffer[80];
 
 	if (!file2str(fn, buffer, sizeof(buffer)))
